move binocular blink ratio test to functional tests
